Fix floating-point type bug

+++ YOLOv4-for-studying/YOLOv4_utils.py
@@ -129,7 +129,7 @@ def image_preprocess(image, target_size, gt_boxes=None, sizex2_flag=False):
     else:
         image_resized = cv2.resize(image, (new_image_w, new_image_h))                     #the original image is resized into 416 x smaller coordinate
 
-    image_padded = np.full(shape=[target_size_h, target_size_w, 3], fill_value=128.0)
+    image_padded = np.full(shape=[target_size_h, target_size_w, 3], fill_value=128.0, dtype=np.float32)
     dw, dh = (target_size_w - new_image_w) // 2, (target_size_h - new_image_h) // 2
     image_padded[dh:new_image_h+dh, dw:new_image_w+dw] = image_resized                #pad the resized image into image_padded
     image_padded = image_padded/255.0
